refactor: move test files into directory structure
