+++ api/app/tests/common/exceptions.py
