+++ api/app/tests/common/ticket_utils.py
@@ -7,7 +7,7 @@ from sqlalchemy.orm import Session
 from app import models
 from app.main import app
 from app.routers.pteams import bg_create_tags_from_sbom_json
-from app.tests.medium.utils import (
+from app.tests.common.utils import (
     create_pteam,
     create_user,
     create_vuln,
