+++ api/app/tests/integrations/business/test_eol_business.py
@@ -9,11 +9,11 @@ from app import models, persistence
 from app.business.eol import eol_business
 from app.notification.eol_notification_utils import EOL_WARNING_THRESHOLD_DAYS
 from app.routers.pteams import bg_create_tags_from_sbom_json
-from app.tests.medium.constants import (
+from app.tests.common.constants import (
     PTEAM1,
     USER1,
 )
-from app.tests.medium.utils import (
+from app.tests.common.utils import (
     create_pteam,
     create_user,
 )
