+++ api/app/tests/integrations/business/test_threat_business.py
@@ -4,10 +4,10 @@ from sqlalchemy.orm import Session
 
 from app import models, persistence
 from app.business import threat_business
-from app.tests.medium.constants import (
+from app.tests.common.constants import (
     USER1,
 )
-from app.tests.medium.utils import (
+from app.tests.common.utils import (
     create_user,
 )
