+++ api/app/tests/integrations/business/test_ticket_business.py
@@ -4,12 +4,12 @@ from sqlalchemy.orm import Session
 
 from app import models, persistence
 from app.business import ticket_business
-from app.tests.medium.constants import (
+from app.tests.common.constants import (
     PTEAM1,
     USER1,
     USER2,
 )
-from app.tests.medium.utils import (
+from app.tests.common.utils import (
     create_pteam,
     create_user,
 )
