+++ api/app/tests/integrations/test_eols.py
@@ -9,11 +9,11 @@ from sqlalchemy.orm import Session
 from app import models
 from app.main import app
 from app.routers.pteams import bg_create_tags_from_sbom_json
-from app.tests.medium.constants import (
+from app.tests.common.constants import (
     PTEAM1,
     USER1,
 )
-from app.tests.medium.utils import (
+from app.tests.common.utils import (
     create_pteam,
     create_user,
     headers_with_api_key,
