+++ api/app/tests/integrations/test_pteams.py
@@ -23,7 +23,7 @@ from app.notification.slack import (
 )
 from app.routers.pteams import bg_create_tags_from_sbom_json
 from app.tests.common import ticket_utils
-from app.tests.medium.constants import (
+from app.tests.common.constants import (
     PTEAM1,
     PTEAM2,
     SAMPLE_SLACK_WEBHOOK_URL,
@@ -32,7 +32,7 @@ from app.tests.medium.constants import (
     VULN2,
     VULN3,
 )
-from app.tests.medium.utils import (
+from app.tests.common.utils import (
     create_pteam,
     create_user,
     create_vuln,
