+++ api/app/tests/integrations/test_users.py
@@ -5,7 +5,7 @@ from sqlalchemy.orm import Session
 
 from app import models, persistence, schemas
 from app.main import app
-from app.tests.medium.constants import (
+from app.tests.common.constants import (
     PTEAM1,
     PTEAM2,
     USER1,
@@ -13,8 +13,8 @@ from app.tests.medium.constants import (
     VULN1,
     VULN2,
 )
-from app.tests.medium.exceptions import HTTPError
-from app.tests.medium.utils import (
+from app.tests.common.exceptions import HTTPError
+from app.tests.common.utils import (
     accept_pteam_invitation,
     create_pteam,
     create_user,
