+++ api/app/tests/integrations/test_vulnerability_detector.py
@@ -2,11 +2,11 @@ from sqlalchemy.orm import Session
 
 from app import models
 from app.detector import vulnerability_detector
-from app.tests.medium.constants import (
+from app.tests.common.constants import (
     USER1,
     VULN1,
 )
-from app.tests.medium.utils import (
+from app.tests.common.utils import (
     create_user,
     create_vuln,
 )
