+++ api/app/tests/requests/pteams/test_pteam_progress.py
@@ -7,8 +7,8 @@ from sqlalchemy.orm import Session
 
 from app import models
 from app.main import app
-from app.tests.medium.constants import PTEAM1, USER1, USER2
-from app.tests.medium.utils import (
+from app.tests.common.constants import PTEAM1, USER1, USER2
+from app.tests.common.utils import (
     create_pteam,
     create_user,
     headers,
