+++ api/app/tests/requests/pteams/test_pteam_tickets.py
@@ -10,15 +10,15 @@ from app import models, persistence
 from app.business import ticket_business
 from app.main import app
 from app.tests.common import ticket_utils
-from app.tests.medium.constants import (
+from app.tests.common.constants import (
     PTEAM1,
     PTEAM2,
     USER1,
     USER2,
     VULN1,
 )
-from app.tests.medium.exceptions import HTTPError
-from app.tests.medium.utils import (
+from app.tests.common.exceptions import HTTPError
+from app.tests.common.utils import (
     accept_pteam_invitation,
     create_pteam,
     create_user,
