+++ api/app/tests/requests/pteams/test_pteam_vuln_ids.py
@@ -6,14 +6,14 @@ from fastapi.testclient import TestClient
 from app import models, persistence
 from app.main import app
 from app.tests.common import ticket_utils
-from app.tests.medium.constants import (
+from app.tests.common.constants import (
     PTEAM1,
     PTEAM2,
     USER1,
     USER2,
     VULN1,
 )
-from app.tests.medium.utils import (
+from app.tests.common.utils import (
     create_pteam,
     create_user,
     headers,
