+++ api/app/tests/requests/test_users.py
@@ -4,9 +4,9 @@ from fastapi.testclient import TestClient
 from app import models, persistence
 from app.constants import ZERO_FILLED_UUID
 from app.main import app
-from app.tests.medium.constants import PTEAM1, PTEAM2, USER1, USER2
-from app.tests.medium.exceptions import HTTPError
-from app.tests.medium.utils import (
+from app.tests.common.constants import PTEAM1, PTEAM2, USER1, USER2
+from app.tests.common.exceptions import HTTPError
+from app.tests.common.utils import (
     create_pteam,
     create_user,
     headers,
