+++ api/app/tests/requests/test_vulns.py
@@ -12,8 +12,8 @@ from sqlalchemy.orm import Session
 from app import models
 from app.main import app
 from app.routers.pteams import bg_create_tags_from_sbom_json
-from app.tests.medium.constants import PTEAM1, USER1, USER2
-from app.tests.medium.utils import (
+from app.tests.common.constants import PTEAM1, USER1, USER2
+from app.tests.common.utils import (
     create_pteam,
     create_user,
     headers,
